+++ src/Fledge/Async/NullCancellation.php
@@ -27,21 +27,28 @@ namespace Fledge\Async;
  */
 final class NullCancellation implements Cancellation
 {
+    use ForbidCloning;
+    use ForbidSerialization;
+
+    #[\Override]
     public function subscribe(\Closure $callback): string
     {
         return "null-cancellation";
     }
 
+    #[\Override]
     public function unsubscribe(string $id): void
     {
         // nothing to do
     }
 
+    #[\Override]
     public function isRequested(): bool
     {
         return false;
     }
 
+    #[\Override]
     public function throwIfRequested(): void
     {
         // nothing to do