+++ src/Fledge/Async/Parallel/Context/ThreadContext.php
@@ -151,7 +151,12 @@ final class ThreadContext extends AbstractContext
             $exited = true;
         })->ignore();
 
-        $this->oid = \getmypid();
+        $pid = \getmypid();
+        if ($pid === false) {
+            throw new ContextException("Failed to determine PID");
+        }
+
+        $this->oid = $pid;
     }
 
     public function receive(?Cancellation $cancellation = null): mixed