+++ src/Fledge/Async/TimeoutCancellation.php
@@ -17,18 +17,17 @@ final readonly class TimeoutCancellation implements Cancellation
     private Cancellation $cancellation;
 
     /**
-     * @param float  $timeout Seconds until cancellation is requested.
+     * @param float $timeout Seconds until cancellation is requested.
      * @param string $message Message for TimeoutException. Default is "Operation timed out".
      */
     public function __construct(float $timeout, string $message = "Operation timed out")
     {
-        $this->cancellation = $source = new Internal\Cancellable;
+        $this->cancellation = $source = new Internal\Cancellable();
 
-        $trace = null; // Defined in case assertions are disabled.
-        \assert((bool) ($trace = \debug_backtrace(0)));
+        \assert((bool) ($trace = \debug_backtrace(\DEBUG_BACKTRACE_IGNORE_ARGS)));
 
-        $this->callbackId = EventLoop::delay($timeout, static function () use ($source, $message, $trace): void {
-            if ($trace) {
+        $this->callbackId = EventLoop::delay($timeout, static function () use (&$trace, $source, $message): void {
+            if ($trace ?? false) {
                 $message .= \sprintf("\r\n%s was created here: %s", self::class, Internal\formatStacktrace($trace));
             } else {
                 $message .= \sprintf(" (Enable assertions for a backtrace of the %s creation)", self::class);
@@ -48,21 +47,25 @@ final readonly class TimeoutCancellation implements Cancellation
         EventLoop::cancel($this->callbackId);
     }
 
+    #[\Override]
     public function subscribe(\Closure $callback): string
     {
         return $this->cancellation->subscribe($callback);
     }
 
+    #[\Override]
     public function unsubscribe(string $id): void
     {
         $this->cancellation->unsubscribe($id);
     }
 
+    #[\Override]
     public function isRequested(): bool
     {
         return $this->cancellation->isRequested();
     }
 
+    #[\Override]
     public function throwIfRequested(): void
     {
         $this->cancellation->throwIfRequested();